Added final changes

+++ src/controllers/tasks.controller.js
@@ -5,7 +5,8 @@ const { isValidObjectId } = require('mongoose');
 class TaskController {
     async getAllTasks(req, res) {
         try {
-            const tasks = await taskService.getAllTasks();
+            let tasks = await taskService.getAllTasks();
+            tasks = tasks.filter((task) => task.createdBy.toString() === req.user._id.toString());
             res.status(200).json({ message: 'Fetched all tasks successfully', success: true, tasks});
         } catch (error) {
             console.error(error);
@@ -41,6 +42,9 @@ class TaskController {
             if (!task) {
                 return res.status(404).json({ message: 'Task not found', success: false });
             }
+            if(task.createdBy.toString() !== req.user?._id.toString()){
+                return res.status(403).json({ message: 'Cannot access Task!', success: false });
+            }
             res.status(200).json({ message: 'Fetched task details successfully', success: true, task });
         } catch (error) {
             console.error(error);
@@ -49,7 +53,7 @@ class TaskController {
     }
 
     async updateTask(req, res) {
-        if(req.params.id){
+        if(!req.params.id){
             return res.status(400).json({ message: 'Task ID is required', success: false });
         }
         if(!isValidObjectId(req.params.id)){
@@ -61,11 +65,15 @@ class TaskController {
         }
         const { title, description, status } = req.body;
         try {
-            const task = await taskService.updateTask(req.params.id, { title, description, status });
+            let task = await taskService.getTaskById(req.params?.id);
+            if(task.createdBy.toString() !== req.user?._id.toString()){
+                return res.status(403).json({ message: 'Cannot update task!', success: false });
+            }
+            task = await taskService.updateTask(req.params?.id, { title, description, status });
             if (!task) {
                 return res.status(500).json({ message: 'Error while updating task', success: false });
             }
-            res.status(201).json({ message: 'Updated task successfully', success: true, task });
+            res.status(201).json({ message: 'Task updated successfully', success: true, task });
         } catch (error) {
             console.error(error);
             res.status(500).json({ message: 'Server Error', success: false });
@@ -83,6 +91,9 @@ class TaskController {
             const task = await taskService.getTaskById(req.params?.id);
             if(!task){
                 return res.status(404).json({ message: 'Task not found', success: false });
+            }
+            if(task.createdBy.toString() !== req.user?._id){
+                return res.status(403).json({ message: 'Cannot delete task', success: false });
             }
             await taskService.deleteTask(task?._id);
             res.json({ message: 'Task deleted', success: true });
